Moved mastermodel related dto and mapping logic
