+++ maestro-webapi/src/main/java/org/intocps/maestro/webapi/maestro2/Maestro2ScenarioController.java
@@ -8,11 +8,11 @@ import core.ScenarioLoader;
 import org.apache.commons.io.FileUtils;
 import org.intocps.maestro.core.messages.ErrorReporter;
 import org.intocps.maestro.framework.fmi2.Fmi2SimulationEnvironmentConfiguration;
-import org.intocps.maestro.webapi.MasterModelMapper;
+import org.intocps.maestro.plugin.MasterModelMapper;
 import org.intocps.maestro.webapi.dto.ExecutableMasterAndMultiModelTDO;
 import org.intocps.maestro.webapi.dto.MasterMultiModelDTO;
 import org.intocps.maestro.webapi.dto.VerificationDTO;
-import org.intocps.maestro.webapi.maestro2.dto.MultiModelScenarioVerifier;
+import org.intocps.maestro.core.dto.MultiModelScenarioVerifier;
 import org.intocps.maestro.webapi.util.Files;
 import org.intocps.maestro.webapi.util.ZipDirectory;
 import org.springframework.http.MediaType;
